BUILD(cmake): Fix usage of set_property
We used this function to set multiple properties at once which is not
supported. Instead what happens is that the first referenced property is
populated with a list of values (the following words in the set_property
function).

Thus the set_property call had to be separated into two individual calls
each setting only a single property to the desired value.