REFAC(client): Add and register stream operators
to the Input structs

Step 1 to fix Windows clients loading / saving shortcuts since 3dc7b00d0919cf2400071deba54ff7f64ee3d9ea

Adds the stream operators to the Input structs and a call to register them in Settings.cpp
The metatypes have to be registered in Settings.cpp because it runs before GlobalShortcuts are initialized. Without being registered it will not load the Input buttons.