REFAC(client): Don't use deprecated mutex ctor
In the latest Qt version, specifying a recrsion mode when creating a
mutex was deprecated in favor of QRecursiveMutex being used for the
cases where recursion is needed.

Given that we only ever needed non-recursive mutexes, all we had to do
was to remove the argument to the ctor.

Ref.: Upstream commit deprecating that ctor:
https://github.com/qt/qtbase/commit/d4b206b246caf9b49110526585693ab629609d99

Fixes #4973