BUILD(cmake): Revamp translation handling
Previously we re-generated our translation files when building, causing
them to change in the source tree.

With this new approach, we assume that the .ts files are already
up-to-date and thus we only compile them to .qm files that are stored in
the build directory. That way the source dir remains uncluttered.