FIX(client, ui): TalkingUI font not scaling properly
When scaling the font size system-wide (e.g. set to 150%), the TalkingUI
might not adapt to this properly.

This started happening with the introduction of the
MultiStyleWidgetWrapper class (commit
ff040fb23c407ee34e6d8796a3f058a455f40cb1). As it turns out the
conversion from pixel to point size performed in its setFontSize did not
work properly, causing an incorrect font size to be used. And as the
font size also determines the icons size, these were not scaled properly
either.

The fix is to simply skip on the pixel to point conversion entirely and
instead use the appropriate unit when setting the respective StyleSheet
property. That way Qt can figure the conversion out as needed.