BUILD(cmake, windows): Prevent definition  min/max macros
These macros are defined in Windows.h if the NOMINMAX macro is not set.
This then causes all other uses of the terms "min" and "max" (as e.g. in
std::min, std::numeric_limits<>::min, etc.) to fail, which is
unacceptable. Therefore NOMINMAX is now defined globally when compiling
on Windows.