FIX(client): Don't escape characters in URL targets
If one was to send a link like
"http://localhost/?var1=foo&var2=bar&var3=baz" either as-is or as a
markdown-style link ("[name](<url>)"), the "&" character in the link
would get escaped as "&amp;".

For displaying purposes this is fine, as the Mumble client re-interprets
these escape sequences before rendering them to the user. The same
escaping was done in the surrounding <a> HTML element's href
specification. This lead to invalid URLs being used as link targets.

Therefore this commit makes sure that the href content of an <a> HTML
element will be un-escaped before inserting into the HTML structure
(thus preserving the original characters in the href specification).

Fixes #4999