CHANGE(client, ui): Display unit as suffix in SpinBox
The units of the settings for positional audio were displayed behind the
SpinBox as an additional label.

This commit removes the labels and instead uses the "suffix" field of
the SpinBoxes to display the unit directly in the SpinBox itself.

In principal this also has the potential to fix the issue of the unit
being placed incorrectly when an right-to-left locale is being used as in
that case Qt should put the suffix on the other side.
Note that this has not been verified though.