FIX(client): Positional audio not working properly after canceling audio wizard
The "bPosTest" variable indicates whether we're in the middle of a test.

It causes the user's direction vector to freeze so that it always "looks" straight ahead.

When completing the wizard, accept() is called and thus the variable is set to false again.

However, when the wizard is closed prematurely, reject() is called instead.