MAINT: Integrate theme submodule
The theme is now integrated into the main repository. This facilitates
future work on it as we don't have to work across two repositories.

The theme is overtaken from https://github.com/mumble-voip/mumble-theme
without any modifications to it. All file from there are copied to this
repo now.

The theme's directory was renamed from "Mumble" to "Default" in order to
a) Make the name more descriptive
b) Avoid conflicts updating already checked out workspaces (git would
report conflicts if the new dir was called the same as the old one)