FIX(audiowizard): Reset AudioOutput ownership before restartAudio()
The AudioOutputPtr ao holds a reference to the audio output. Calling
wipe() only clears the buffers, but the reference still exists.

When we call restartAudio() without ao going out of scope, then
Audio::stop() will loop endlessly as we are waiting for all audio output
references to clear.

We fix this deadlock by explicitly resetting the shared_ptr ao, such
that it doesn't hold a reference anymore when restartAudio() is called.

Fixes #5081