MAINT: Translations backport script
Up to now when a stable release series was split off from the master
branch, there was essentially no (good) way of backporting new
translations for source strings in the stable series.

This is because usually the master branch contains new or changed source
strings and thus simply merging the respective commits into the stable
series can cause all sorts of trouble.

The script added by this commit takes care of that by comparing
translation files translation by translation and only updates the
translation if there is a newer translation for the exact same source
string in the target (master) branch.

This takes care of cases in which translations are added or altered
after the stable series has been split off.
It does not however work as soon as the source string for that
translation has changed as that will cause the script to consider these
as different strings and thus no updating will be performed.

Fixes #4740