FIX(client): Differentiate plugins for different OS
Using which library can be loaded on the current OS is not a sufficient
criterion for differentiating between plugin libraries for different OS
as there is some overlap in supported file formats (e.g. .so is
supported on Linux and macOS).

Instead the new approach relies on suffixes in the plugin library's name
in order to determine which one to load.