FIX(client): PluginInstaller unable to extract zips
On Windows the PluginInstaller was unable to extract certain zip files.

As it turns out the issue was that the filestreams were not opened
in binary mode which threw Poco off in certain situations.

Therefore the fix simply is to use binary mode when working with
these file streams. Note that the used Poco filestream class operates
in binary mode by default.