FIX(server): uiSession being used before assignment
In eabfbab403e6ba10219c262518899db5a60d0c1e the assignment of the
uiSession variable of any given client was delayed into the
Server::msgAuthenticated function in order to prevent non-Mumble
connections to be counted as users.

However the assignment of the session ID was moved so far down that
there is now previous code that already depends on its value and is now
using an uninitialized (or at least wrong) value for the session ID.

This commit fixes this by assigning the session ID very early in the
Server::msgAuthenticated function, making sure that the code before it
does not reference the session ID already.

This bug could have had all kinds of side-effects but one of them was
that users were always presented with the "Unable to find matching CELT
codecs with other clients." message upon connecting to the server.

Fixes #5051