FEAT(positional-audio): Add support for DT_GNU_HASH, to locate exported symbols
Our Source Engine plugin retrieves the interfaces on Linux through "s_pInterfaceRegs", an exported symbol.

Left 4 Dead 2 just received a big update and the binaries don't have the DT_HASH table anymore.
As a result, the plugin was not finding the symbol anymore.

This is basically a follow-up to 6f19d7ebfd7565843b3c56484af624afb5956c0f, which implemented support for DT_GNU_HASH in the overlay code.