FIX(server): Value queried but not used in DB insertion
Inside ServerDB there was one SQL statement that reads values off from
one table and writes them into another one. However while doing so, it
(in contrast to the other statements doing the same for other DB
versions) also queries the `last_disconnect` value. This values is
unused in the INSERT statement though and therefore reading it in the
first place seems to be pointless.