FEAT(client): Introduce Positional Audio Viewer, replacement for MumblePAHelper
MumblePAHelper (Mumble Positional Audio Helper) is the tool we had been using to test positional audio plugins.

As the application is basically just a dialog, we thought many times in the past about merging it in Mumble.

Now that MumblePAHelper is obsolete (it doesn't support the new API), it's the perfect time to do that.

The dialog can be opened through a new action in the "Developer" menu.

Showcase: https://odysee.com/positional-audio-viewer-left4dead:f4ad40f7e55797e6c8d2b1817915cab71b4893f4