FIX(client): Duplicate name in Log.ui
The UI file was using a name that was alredy in use. This has been
introduced with 6a10b4627fab7352601e4bde8d0a440c132bf7b3. However as of
now this did not cause any trouble except a warning during building.

Nonetheless this commit fixes the duplicate name. In fact the fix is
performed by changing how the "problematic" component is laid out in
the first place (entirely removing the need for a GridLayout).

Fixes #5148