FIX(client): Plugin updater keeping plugin locked
PR #5152 overhauled the code for the plugin installer such that it
can completely unload a plugin before attempting to overwrite it.

However when the installer is called from the updater, that doesn't work
since the installer itself was still holding a handle to that plugin,
preventing it from unloading. Therefore this commit makes sure that the
installer releases its handle before calling the installer.

Fixes #4946