FIX(client, audio): PulseAudio not initializing
PR #5159 removed a few seemingly unneeded waits on audio system
initialization which drastically sped up Mumble's startup. However as it
turns out for PulseAudio some waiting is actually required as the system
initializes asynchronously but we require the knowledge about the init
result.

Instead of adding an arbitrary wait time again though, this commit adds
a mechanism that waits exactly as long as necessary for the init to be
completed and then continues. That still cuts down the waiting time
tremendously when comparing with the previous (arbitrary) wait time of 1
second.

Fixes #5181