CI(azure): Don't package AppImage
The packaging of the AppImage did not yield usable binaries for a long
time now but it does take up a significant amount of time in the Linux
CI and (more importantly) tends to fail every now and then causing the
entire build to fail with it.

For this reason, we will no longer bundle an AppImage at least until we
properly fixed the underlying issues and are able to produce working
AppImages again.