FIX(installer): Generate output filename from release ID
Right now the output filename looks like this:

- mumble_client-1.4.0.224-x64.msi
- mumble_server-1.4.0.224-x64.msi

We had always been renaming the binaries manually before signing and uploading them on our server.

Now that we use an automatic signing workflow with AppVeyor + SignPath, the undesired initial filename is an issue.
The package is identified as "{FILENAME} by github.com/mumble-voip/mumble", meaning that the final result would be:

- mumble_client-1.4.0.224-x64.msi by github.com/mumble-voip/mumble
- mumble_server-1.4.0.224-x64.msi by github.com/mumble-voip/mumble

This commit fixes the installer builder so that it generates the filename based on the release ID:

- mumble-client_1.4.0-rc1.x64.msi
- mumble-server_1.4.0-rc1.x64.msi

Extra changes:

- The server's product name was "Mumble Server". It's now "Mumble (server)", consistent with "Mumble (client)".
- Redundant version assignments removed from Main(). ClientInstaller() and ServerInstaller() take care of that.