FIX(client): Don't prepend plugin name to context
With the change to the new plugin system a change was introduced that
would prepend the current plugin's name to its context in order to
avoid context clashes between plugins.

As it turns out though, this breaks intercompatibility between the 1.4.x
series and all previous versions.

Therefore this commit removes this logic again.

Fixes #5217