BUILD(opus): Fetch submodule from upstream repository
Upon switching to CMake, we kept using our own fork for the repository because a few fixes for the build were needed.

Now that the upstream repository works correctly as submodule, we can fetch from it directly.

Please note that tests provided in the Opus repository are not built anymore.

Fixes #5223.