FIX(client): Send UDP pings only when TCP isn't forced
Previously Mumble would continue answering to the server's UDP pings, even if
TCP mode was enforced. This, in turn, would trick the server into thinking that the client
is still using UDP and continue sending UDP packets to the client.

This commit makes sure that once TCP mode is enabled, the client no longer answers
to UDP pings, causing the server to fall back to TCP as well.

Fixes #4881