FIX(client): Crash when starting recording
If the used audio output system was bogus and therefore configured to
use a 0Hz sample rate, starting a recording would crash Mumble due to an
assertion in the VoiceRecorder class.

This commit adds a check for this situation and instead of crashing
Mumble, this will log a warning message to the console and close the
VoiceRecorderDialog instead.