CHANGE(client, ui): Recording disabled by default
The recording button is now disabled by default. It gets enabled only
when connected to a server (that allows recording). This is done because
the recording feature can't be used without a server connection anyway
(doing so will result in an error popup). Thus this restriction is now
reflected in the UI as well.