BUILD(client): Fix compile errors without manual plugin
When excluding the manual plugin from the build (-Dmanual-plugin=OFF),
the compilation would fail due to the ManualPlugin.h header being
included in certain files nonetheless.

As it turns out the include was superfluous in the PluginConfig.cpp
file and in PluginManager.cpp the include is now shielded by a proper
ifdef-guard.

Fixes #5304