MAINT: Add GSL as submodule
The "Guidelines Support Library" (GSL) is a library that implements
functions and types suggested by the C++ Core Guidelines. This should
make it relatively straight forward to follow these guidelines and thus
to write better C++ code.

This explicitly adds the Microsoft GSL as that seems to be the most
popular (and complete) GSL implementation available.