FIX(client): DB transactions not effective
In the DB code at some places several SQL statements where packed in a
transaction to ensure atomicity of these instructions. However, instead
of accessing our created DB instance directly, they started the
transaction on the DB returned by the static function
QSqlDatabase::database(), which returns a DB with a default name. Since
we are using an explicit name for our DB, this default DB does not exist
and therefore starting the transaction on it will do nothing.

This commit fixes this by starting and committing the transaction on our
created DB instance directly.