REFAC(client): Don't delete shortcuts on save
The Database::setShortcut function is used to save server-specific
shortcuts to the DB. However, at the same time it also deletes these
shortcuts from the passed list. This is a very unintuitive side-effect
of this function that also prevents saving shortcuts intermediately
instead of only upon disconnecting from a server (at which point we
indeed want to remove server-specific shortcuts).

Therefore, this commit removes this side-effect and instead explicitly
removes these shortcuts on disconnecting from a server.