CHANGE(client): Display lost packages instead of received ones
In the server browser, the tooltip for the different entries shows the
packet loss. It is displayed as a percentage and as a x/y stat. The
latter currently displays <received>/<sent> which is very confusing,
given that the title for that row is "Packet loss" implying x would
describe the lost packages instead of the received ones.

Thus this commit changes this to the expected behavior. Furthermore it
makes sure that <lost> is always <= <sent>.