BUILD(client): Add missing ifdef guards
0e17c5394c5597b746c6076d16c732e2a4a04e7f introduced a migration system
from migrating from the old shortcut format to the newer one (on
Windows) but it seems in this new function it was forgotten to check
whether support for GKey and/or XBox-Input was actually included or not.

This could lead to compile errors when e.g. using -Dgkey=OFF when
invoking cmake.

This commit fixes that by adding the mentioned include-guards.

Fixes #5322