FIX(client): Duplicate plugin entries
Before when installing Mumble (on Windows) it could happen that when
viewing the plugin settings, every plugin would be listed multiple
times. This was caused by duplicate entries in the plugins search path
used by Mumble. The duplicates arose from using non-canonical paths in
the set for storing the search paths which allowed to add the same
directory twice by describing its path slightly differently.

This commit makes sure that all paths are first canonicalized before
they are added to the set, such that only distinct paths are maintained.

Fixes #5271