FIX(client): Error if API call can't work due to server
The sendData API call requires the server to be of version 1.4.0 or
newer to work. Thus, when calling this API function while connected to
an older server, we can already tell that this won't work. Previously
this was silently ignored.

This commit introduces a new API error that will inform the user about
this problem. This should clear up the confusion as to why the sendData
call is apparently not doing anything.

Fixes #5331

Co-authored-by: Davide Beatrici <github@davidebeatrici.dev>