FIX(client): Crash on plugin link/unlink
When a positional data plugin lost link while Mumble was playing audio
(or recording & sending audio from the mic), it could end up crashing.

The reason for this was that whenever a plugin lost link,
PluginManager::reportLostLink was called. In this function a
corresponding message was logged to Mumble's console. This involves
accessing a Qt GUI object. However, if the lost link was detected from
within the audio thread (either AudioInput or AudioOutput) this function
would be called from a different thread than the GUI (main) thread. This
results in undefined behavior and in our case causes a segmentation
fault, which crashes Mumble.

Mitigating this issue is a simple matter of converting this function
into a slot and then only emitting the corresponding signal instead of
calling the function directly. This way, the lostLink function is only
ever called from within the plugin manager's thread (which lives in the
GUI thread).

The same argument applies to the messages logged upon linking a plugin
or upon reporting that a plugin has encountered a permanent error.

Fixes #5319