BUILD: Use build number instead of patch version
Using a four-segment version (major.minor.patch.build) is causing issues
on Windows, because it tends to ignore the fourth version number in most
cases. This can cause upgrades to fail and/or multiple entries being
created in the installed applications list.

For this reason, this commit changes the version format to be
major.minor.build.

Fixes #4921