BUILD: Fix violation of ODR
The "ODR" (One Definition Rule) states that any function/object/etc. may
only be defined once. However, since we are using some preprocessor
magic to only include certain fields in the Channel struct, we have to
make sure that everywhere Channel.h is included, these macros are set
accordingly.

However, the User.h was part of the "shared" library and in that case no
macros were being set when compiling this translation unit. In the
actual client however, the "MUMBLE" macro is set, such that all headers
being part of the "mumble" target will use a different Channel
definition than the User class which is part of the "shared" target.

In order to not violate the ODR, the User class was not excluded from
being part of the "shared" target and is instead explicitly included in
the "mumble" and "mumble-server" targets (making sure the necessary
macros will be set).